// src/types/types/auth.types.ts
export interface RcDetails{
    id_number: string
}

export interface ApiErrorResponse {
    status: boolean;
    message: string;
    statuscode: number;
}

export interface RCApiResponse {
    reference_id: number;
    statuscode: number;
    message: string;
    status: boolean;
    data: RCData;
}

export interface RCData {
    rc_number: string;
    fit_up_to: string;
    registration_date: string;
    owner_name: string;
    father_name: string;
    present_address: string;
    permanent_address: string;
    mobile_number: string;
    vehicle_category: string;
    vehicle_chasi_number: string;
    vehicle_engine_number: string;
    maker_description: string;
    maker_model: string;
    body_type: string;
    fuel_type: string;
    color: string;
    norms_type: string;
    financer: string;
    financed: string;
    insurance_company: string;
    insurance_policy_number: string;
    insurance_upto: string;
    manufacturing_date: string;
    manufacturing_date_formatted: string;
    registered_at: string;
    latest_by: string;
    less_info: boolean;
    tax_upto: string;
    tax_paid_upto: string;
    cubic_capacity: string;
    vehicle_gross_weight: string;
    no_cylinders: string;
    seat_capacity: string;
    sleeper_capacity: string;
    standing_capacity: string;
    wheelbase: string;
    unladen_weight: string;
    vehicle_category_description: string;
    pucc_number: string;
    pucc_upto: string;
    permit_number: string;
    permit_issue_date: string;
    permit_valid_from: string;
    permit_valid_upto: string;
    permit_type: string;
    national_permit_number: string;
    national_permit_issue_date: string;
    national_permit_upto: string;
    national_permit_issued_by: string;
    non_use_status: string;
    non_use_from: string;
    non_use_to: string;
    blacklist_status: string;
    noc_details: string;
    owner_number: string;
    rc_status: string;
    masked_name: boolean;
    challan_details: null | unknown;
    variant: string;
}